platform/x86: serial-multi-instantiate: allow single GpioInt IRQ for INT3515
On some devices, such as the Lenovo ThinkPad T14 Gen1 (AMD), there is
only one GpioInt resource defined for all i2c device instances. Handle
this case appropriately by autodetecting the irq type and allowing
fallback to the first IRQ index for the second, third and fourth tps6598x
instances.

Additionally, to use the `platform_get_irq_optional` function to silence
errors that may not be relevant if the IRQ is optional. In cases where the
IRQ is not optional, `dev_err_probe` is still triggered, so other devices
will not be affected by this change.

Signed-off-by: Michał Kopeć <michal@nozomi.space>